requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.3
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-multiupload==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
MarkupSafe==2.1.3
Pillow==10.0.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
template==0.7.6
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
